[TAN-8601] Move CustomPageHeader out of the show page's folder
