+++ front/app/components/CustomPageHeader/TwoRowLayout.tsx
@@ -5,6 +5,7 @@ import { Box, useBreakpoint } from '@citizenlab/cl2-component-library';
 import { ICustomPageData } from 'api/custom_pages/types';
 
 import { homepageBannerLayoutHeights } from 'containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/HeaderImageDropzone';
+import AdminCustomPageEditButton from 'containers/CustomPageShow/AdminCustomPageEditButton';
 
 import ContentContainer from 'components/ContentContainer';
 import {
@@ -12,7 +13,6 @@ import {
   HeaderImage,
 } from 'components/LandingPages/citizen/TwoRowLayout';
 
-import AdminCustomPageEditButton from './AdminCustomPageEditButton';
 import HeaderContent from './HeaderContent';
 
 interface Props {
